appliance: move yaml package here
It's not particularly generic.
