+++ internal/appliance/reconciler/BUILD.bazel
@@ -84,8 +84,8 @@ go_test(
     },
     deps = [
         "//internal/appliance/config",
+        "//internal/appliance/yaml",
         "//internal/slices",
-        "//internal/yaml",
         "@com_github_go_logr_stdr//:stdr",
         "@com_github_stretchr_testify//require",
         "@com_github_stretchr_testify//suite",
